+++ src/main/java/eu/pb4/mrpackserver/util/Utils.java
@@ -13,7 +13,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
+import java.lang.instrument.Instrumentation;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
 import java.net.URI;
@@ -28,27 +28,55 @@ import java.nio.file.Path;
 import java.security.DigestInputStream;
 import java.security.MessageDigest;
 import java.util.*;
-import java.util.function.Consumer;
 import java.util.jar.Attributes;
 import java.util.jar.JarFile;
 
 public interface Utils {
     Gson GSON = new GsonBuilder().disableHtmlEscaping().setLenient().create();
 
-
-    static void start(Path path, String[] args) throws Throwable {
+    static boolean launchJar(Path path, String... args) {
         String mainClass;
+        String launcherAgentClass = null;
+
         try (var jarFile = new JarFile(path.toFile())) {
             mainClass = jarFile.getManifest().getMainAttributes().getValue(Attributes.Name.MAIN_CLASS);
+            launcherAgentClass = jarFile.getManifest().getMainAttributes().getValue("Launcher-Agent-Class");
             if (mainClass == null) {
-                throw new RuntimeException("No main class!");
+                Logger.error("The '%s' is missing main class!!", path);
+                return false;
             }
+        } catch (Throwable e) {
+            Logger.error("Failed to read data of '%s'!", e);
+            return false;
         }
 
-        //noinspection resource
-        var launchClassLoader = new URLClassLoader(new URL[]{path.toUri().toURL()});
-        var handle = MethodHandles.publicLookup().findStatic(launchClassLoader.loadClass(mainClass), "main", MethodType.methodType(void.class, String[].class));
-        handle.invokeExact(args);
+        try (var launchClassLoader = new URLClassLoader(new URL[]{ path.toUri().toURL() })) {
+            if (launcherAgentClass != null) {
+                if (InstrumentationCatcher.exists()) {
+                    try {
+                        var m = launchClassLoader.loadClass(launcherAgentClass).getDeclaredMethod("agentmain", String.class, Instrumentation.class);
+                        m.setAccessible(true);
+                        var handle = MethodHandles.lookup().unreflect(m);
+                        handle.invoke("", InstrumentationCatcher.get());
+                    } catch (Throwable e) {
+                        Logger.error("Error occurred while invoking Launcher-Agent-Class!", e);
+                    }
+                } else {
+                    Logger.error("The executed jar requires Instrumentation, but it's not set up! This might be related to your launch arguments / java version!");
+                }
+            }
+
+            try {
+                var handle = MethodHandles.publicLookup().findStatic(launchClassLoader.loadClass(mainClass), "main", MethodType.methodType(void.class, String[].class));
+                handle.invoke((Object) args);
+                return true;
+            } catch (SecurityException e) {
+                return e.getMessage() != null && e.getMessage().equals("Success");
+            }
+        } catch (Throwable e) {
+            Logger.error("Exception occurred while executing '%s' with arguments '%s'!", path, String.join(" ", args), e);
+            return false;
+        }
     }
 
     @Nullable
@@ -94,7 +122,7 @@ public interface Utils {
         return null;
     }
 
-    static InstanceInfo checkAndSetupModpack(ModpackInfo modpackInfo, InstanceInfo instance, Path currentDir, Path instanceDataDir) throws Throwable {
+    static InstallResult checkAndSetupModpack(ModpackInfo modpackInfo, InstanceInfo instance, Path currentDir, Path instanceDataDir) throws Throwable {
         var mrpackFile = getMrPackFile(modpackInfo, instanceDataDir);
         if (mrpackFile == null) {
             return null;
@@ -113,8 +141,6 @@ public interface Utils {
                     }
                 }
             }
-
-
             var hashPath = instanceDataDir.resolve("hashes.json");
             var hashes = new HashMap<String, String>();
             if (Files.exists(hashPath)) {
@@ -124,13 +150,19 @@ public interface Utils {
             whitelistedDomains.addAll(Constants.WHITELISTED_URLS);
             whitelistedDomains.addAll(modpackInfo.whitelistedDomains);
 
+            Logger.info("Starting %s of %s (%s)", hashes.isEmpty() ? "installation" : "update", modpackInfo.getDisplayName(), modpackInfo.getDisplayVersion());
+
             var handler = new MrPackInstaller(zip.getPath(""), index, currentDir, instance, hashes, whitelistedDomains);
             handler.prepareFolders();
             handler.cleanupOutdatedFiles();
             var x = new FileDownloader();
             handler.requestDownloads(x);
-            x.downloadFiles(handler.getHashes());
-            handler.extractFolders();
+            var failed = x.downloadFiles(handler.getHashes());
+            if (!failed.isEmpty()) {
+                Logger.error("Failed to download provided files: \n- ", String.join("\n- ", failed));
+                return null;
+            }
+            handler.extractIncluded();
 
             Files.deleteIfExists(hashPath);
             Files.writeString(hashPath, Utils.GSON.toJson(handler.getHashes()));
@@ -141,7 +173,31 @@ public interface Utils {
             newInstance.runnablePath = Objects.requireNonNullElse(handler.getNewLauncher(), instance.runnablePath);
             newInstance.dependencies.putAll(index.dependencies);
 
-            return newInstance;
+            if (handler.getForgeInstaller() != null) {
+                return new InstallResult(newInstance, () -> {
+                    Logger.info("Running installer! Older (Neo)Forge installers will quit after success, requiring you to start the server again!");
+                    try {
+                        Thread.sleep(2000);
+                    } catch (InterruptedException ignored) {
+
+                    }
+                    var t = new Thread(() -> {
+                        for (var i = 0; i < 10; i ++) {
+                            System.out.println();
+                        }
+                        Logger.info("Installer finished, but forces the mrpack4server to exit! Start the server again to run it!");
+                        Logger.info("You should ignore instructions about deleting installer files, as it is not needed!");
+
+                    });
+                    Runtime.getRuntime().addShutdownHook(t);
+                    if (!launchJar(currentDir.resolve(handler.getForgeInstaller()), "--installServer")) {
+                        Logger.warn("Failed to execute the installer! See errors above.");
+                    }
+                    Runtime.getRuntime().removeShutdownHook(t);
+                });
+            }
+
+            return new InstallResult(newInstance, () -> {});
         }
     }
 
@@ -158,12 +214,15 @@ public interface Utils {
             return modpackFile;
         }
         URI uri;
-        long size = -1;
-        String hash = null;
+        long size;
+        String hash;
         if (modpackInfo.url != null) {
             uri = modpackInfo.url;
+            size = modpackInfo.size != null ? modpackInfo.size : -1;
+            hash = modpackInfo.sha512;
         } else {
-            var result = ModrinthModpackLookup.find(modpackInfo.projectId, modpackInfo.versionId, modpackInfo.getDisplayName(), modpackInfo.getDisplayVersion());
+            var result = ModrinthModpackLookup.findVersion(modpackInfo.projectId, modpackInfo.getDisplayName(), modpackInfo.getDisplayVersion(), true,
+                    x -> x.versionNumber.equals(modpackInfo.versionId) || x.id.equals(modpackInfo.versionId));
             if (result == null) {
                 return null;
             }
@@ -183,43 +242,41 @@ public interface Utils {
             Logger.error("Failed to locate source mrpack file!", e);
         }
 
-
-
         return null;
     }
 
 
 
-    static byte @Nullable [] handleDownloadedFile(Path out, InputStream body, String path, long fileSize, @Nullable String sha512) {
+    static byte @Nullable [] handleDownloadedFile(Path out, InputStream body, String displayName, long fileSize, @Nullable String sha512) {
         try (var file = Files.newOutputStream(out)) {
             var stream = new DigestInputStream(body, MessageDigest.getInstance("SHA-512"));
             stream.on(true);
-            Logger.info("Downloading file '%s': 0%%", path);
+            Logger.info("Downloading file '%s': 0%%", displayName);
             long current = 0;
             var startingTime = System.currentTimeMillis();
 
             while (true) {
                 var dat = stream.readNBytes(Math.max(stream.available(), 512));
-                if (dat.length == 0) {
+                if (dat == null || dat.length == 0 ) {
                     break;
                 }
                 current += dat.length;
                 file.write(dat);
 
                 if ((System.currentTimeMillis() - startingTime) % 2000 == 1000) {
-                    Logger.info("Downloading file '%s': %s%%", path, fileSize != -1 ? String.valueOf(current * 100 / fileSize) : '?');
+                    Logger.info("Downloading file '%s': %s%%", displayName, fileSize != -1 ? String.valueOf(current * 100 / fileSize) : '?');
                 }
             }
             var hash = stream.getMessageDigest().digest();
             if (sha512 == null || Arrays.equals(HexFormat.of().parseHex(sha512), hash)) {
-                Logger.info("Finished downloading file '%s' successfully!", path);
+                Logger.info("Finished downloading file '%s' successfully!", displayName);
                 return hash;
             } else {
-                Logger.warn("Couldn't validate file '%s'! Expected hash: %s, got: %s", path, sha512, HexFormat.of().formatHex(hash));
+                Logger.warn("Couldn't validate file '%s'! Expected hash: %s, got: %s", displayName, sha512, HexFormat.of().formatHex(hash));
                 return null;
             }
         } catch (Throwable e) {
-            Logger.error("Couldn't download file '%s' correctly!", path, e);
+            Logger.error("Couldn't download file '%s' correctly!", displayName, e);
             return null;
         }
     }
@@ -230,8 +287,10 @@ public interface Utils {
 
     static HttpRequest createGetRequest(URI uri) {
         return HttpRequest.newBuilder(uri)
-                .setHeader("User-Agent", Constants.USER_AGENT)
+                .setHeader("User-InstrumentationCatcher", Constants.USER_AGENT)
                 .GET()
                 .build();
     }
+
+    record InstallResult(InstanceInfo info, Runnable installer) {}
 }
